Improve AM_GEN rules
Get rid of useless extra shell invocations, make sure the rules
fail on error, and remove some quoting as well.

The quoting would be nice to keep around, but since the
gobject-introspection rules already break horribly if eg. any
whitespace is present in $(srcdir), might as well go with it.