Add Resources section to README
It's been pointed out that GitHub considers every tag in a
repository to be a "release", and automatically lists it in
a specific section on the project's page.

Additionaly, for each such "release", it provide links that
can be used to download the contents of the repository at the
time the tag was created.

However, those contents are neither enough to build Cattle
directly (generated files have, rightfully, never been checked
into Git) nor to initialize the build system properly, which
makes the archives created this way all but useless.

Include a new section in the README that explains where to get
proper release archives, plus other information about
obtaining Cattle.

At the same time get rid of the Contacts section, which was
never very useful.