Update and tweak documentation
Make sure documents all have titles so that they are rendered
nicely into HTML, reflow them, and add a few extra bits of markup
where it makes sense.

INSTALL.rst is significantly expanded, and now contains not only
the list of requirements but also actual information on how to
build and install the project.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>